Move P2P snocket utils to Ouroboros.Network.Diffusion.Utils module
